+++ test/app/components/too_many_sidecar_files_component.rb
@@ -1,6 +1,4 @@
 # frozen_string_literal: true
 
 class TooManySidecarFilesComponent < ViewComponent::Base
-  def initialize(*)
-  end
 end