+++ CHANGELOG.md
@@ -4,6 +4,10 @@
 
     *Joel Hawksley*
 
+* Remove initializer requirement from the component.
+
+    *Vasiliy Ermolovich*
+
 # v2.1.0
 
 * Support rendering collections (e.g., `render(MyComponent.with_collection(@items))`).